Flask/requirements.txt
alembic==1.13.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.23.1
bleach==5.0.1
blinker==1.7.0
cffi==1.15.1
click==8.1.7
colorama==0.4.6
colorzero==2.0
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.13
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gpiozero==2.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.13.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Mako==1.3.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.9
psycopg2-binary==2.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.9.0
python-json-logger==2.0.4
python-socketio==5.11.2
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.2
style==1.1.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
typing_extensions==4.11.0
tzdata==2024.1
update==0.0.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.2
widgetsnbextension==4.0.5
wsproto==1.2.0
zipp==3.11.0
